fix: apply review hints
